Revert "WSLg: update FreeRDP working branch to 2.10.0 (#4)"
This reverts commit 74f59827764447dbe67417fd4c2d35893a8e35a7.

+++ winpr/libwinpr/registry/registry_reg.h
@@ -30,7 +30,7 @@ struct _reg
 	FILE* fp;
 	char* line;
 	char* next_line;
-	size_t line_length;
+	int line_length;
 	char* buffer;
 	char* filename;
 	BOOL read_only;
@@ -44,8 +44,7 @@ struct _reg_val
 	RegVal* prev;
 	RegVal* next;
 
-	union reg_data
-	{
+	union reg_data {
 		DWORD dword;
 		char* string;
 	} data;
